Changed the definitions file in the roles/sway/files/etc/sway directory. Updated the value of $rofi_theme to use specific color variables instead of generic ones.
Changed the main.yml file in the roles/sway/tasks directory. Removed the --backup-dir=.backup option from the rsync command.

Changed the Archlinux.yml file in the roles/sway/vars directory. Added wofi to the list of packages.

Changed the main.yml file in the roles/terminal/tasks directory. Replaced the list of terminal packages with the terminal.packages variable.

Changed the main.yml file in the roles/user/tasks directory. Added a task to install the user-dirs.dirs file and a task to sync the /etc/skel directory with each user’s home directory.